Internal refactor of cron
This was performed to prepare for #219. No new functionality was
implemented as part of this commit.
